requirements.txt
# ВКР: Advanced RAG-система
# Работает и локально, и на Streamlit Community Cloud

requests>=2.31.0
lxml>=5.0.0
sentence-transformers>=2.7.0
torch>=2.1.0
numpy>=1.26.0
chromadb>=0.5.0
rank-bm25>=0.2.2
gigachat>=0.1.35
networkx>=3.2
streamlit>=1.33.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
pytest>=8.0.0
